requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-autoslug==1.9.8
django-bootstrap4==2.3.1
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-mathfilters==1.0.0
django-precise-bbcode==1.2.13
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
easy-thumbnails==2.7.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==2.1.1
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.3.0
zipp==3.4.1
